create libs/authenticate.py and move authenticate functions to this file
