+++ src/libs/authenticate.py
@@ -1,0 +1,66 @@
+from datetime import datetime, timedelta
+from fastapi import Depends, HTTPException, status
+from fastapi.security import OAuth2PasswordBearer
+from sqlalchemy.ext.asyncio import AsyncSession
+from passlib.context import CryptContext
+from jose import JWTError, jwt
+from src.db import get_db
+from src.schema import user as user_schema
+from src.cruds import user as user_crud
+
+SECRET_KEY = "09d25e094faa6ca2556c818166b7a9563b93f7099f6f0f4caa6cf63b88e8d3e7"
+ALGORITHM = "HS256"
+ACCESS_TOKEN_EXPIRE_MINUTES = 30
+
+pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
+
+oauth2_schema: OAuth2PasswordBearer = OAuth2PasswordBearer(tokenUrl="/token")
+
+async def get_current_user(
+    token: str = Depends(oauth2_schema),
+    db: AsyncSession = Depends(get_db)
+):
+    credentials_exception = HTTPException(
+        status_code=status.HTTP_401_UNAUTHORIZED,
+        detail="could not validate credentials",
+        headers={"WWW-Authenticate": "Bearer"}
+    )
+    try:
+        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
+        username: str = payload.get("sub")
+        if username is None:
+            raise credentials_exception
+        token_data = user_schema.TokenData(username=username)
+    except JWTError:
+        raise credentials_exception
+    user = await user_crud.get_user(db, username=token_data.username)
+    if user is None:
+        raise credentials_exception
+    return user
+
+async def get_current_active_user(current_user: user_schema.User = Depends(get_current_user)):
+    if not current_user.activated:
+        raise HTTPException(status_code=400, detail="Inactive user")
+    return current_user
+
+def get_hashed_password(password) -> str:
+    return pwd_context.hash(password)
+
+def verify_password(plain_password: str, hashed_password: str) -> bool:
+    return pwd_context.verify(plain_password, hashed_password)
+
+async def authenticate_user(db: AsyncSession, username: str, password: str):
+    user = await user_crud.get_user(db, username)
+    if not user:
+        return False
+    if not verify_password(password, user.password):
+        return False
+    return user
+
+def create_access_token(username: str):
+    expire = datetime.now() + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
+    payload = {
+        "sub": username,
+        "exp": expire
+    }
+    return jwt.encode(payload, SECRET_KEY, algorithm=ALGORITHM)
