+++ src/routers/user.py
@@ -1,102 +1,22 @@
-from datetime import datetime, timedelta
 from typing import List
 from fastapi import APIRouter, Depends, HTTPException, status
-from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm
+from fastapi.security import OAuth2PasswordRequestForm
 from sqlalchemy.ext.asyncio import AsyncSession
-from sqlalchemy import select
-from sqlalchemy.engine import Result
-from jose import JWTError, jwt
-from passlib.context import CryptContext
-from src.models import model
 from src.schema import user as user_schema, task as task_schema
 from src.cruds import user as user_crud
+from src.libs import authenticate
 from src.db import get_db
 
-SECRET_KEY = "09d25e094faa6ca2556c818166b7a9563b93f7099f6f0f4caa6cf63b88e8d3e7"
-ALGORITHM = "HS256"
-ACCESS_TOKEN_EXPIRE_MINUTES = 30
-
-fake_users_db = {
-    "johndoe": {
-        "id": 1,
-        "username": "johndoe",
-        "full_name": "John Doe",
-        "email": "johndoe@example.com",
-        "password": "$2b$12$EixZaYVK1fsbw1ZfbX3OXePaWxn96p36WQoeG6Lruj3vjPGga31lW",
-        "activated": True,
-    }
-}
-class UserInDB(user_schema.User):
-    password: str
-
-pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
-
-oauth2_schema: OAuth2PasswordBearer = OAuth2PasswordBearer(tokenUrl="/token")
-
 router = APIRouter(prefix="", tags=["user"])
-
-async def get_user(db: AsyncSession, username: str):
-    stmt = select(model.User).where(model.User.username == username)
-    result: Result = await db.execute(stmt)
-    return result.scalar()
-
-async def get_current_user(
-    token: str = Depends(oauth2_schema),
-    db: AsyncSession = Depends(get_db)
-):
-    credentials_exception = HTTPException(
-        status_code=status.HTTP_401_UNAUTHORIZED,
-        detail="could not validate credentials",
-        headers={"WWW-Authenticate": "Bearer"}
-    )
-    try:
-        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
-        username: str = payload.get("sub")
-        if username is None:
-            raise credentials_exception
-        token_data = user_schema.TokenData(username=username)
-    except JWTError:
-        raise credentials_exception
-    user = await get_user(db, username=token_data.username)
-    if user is None:
-        raise credentials_exception
-    return user
-
-async def get_current_active_user(current_user: user_schema.User = Depends(get_current_user)):
-    if not current_user.activated:
-        raise HTTPException(status_code=400, detail="Inactive user")
-    return current_user
-
-def get_hashed_password(password) -> str:
-    return pwd_context.hash(password)
-
-def verify_password(plain_password: str, hashed_password: str) -> bool:
-    return pwd_context.verify(plain_password, hashed_password)
-
-async def authenticate_user(db: AsyncSession, username: str, password: str):
-    user = await get_user(db, username)
-    if not user:
-        return False
-    if not verify_password(password, user.password):
-        return False
-    return user
-
-def create_access_token(username: str):
-    expire = datetime.now() + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
-    payload = {
-        "sub": username,
-        "exp": expire
-    }
-    return jwt.encode(payload, SECRET_KEY, algorithm=ALGORITHM)
 
 @router.post("/user", response_model=user_schema.Token)
 async def register_user(
     form_data: user_schema.UserCreate = Depends(),
     db: AsyncSession = Depends(get_db)
 ):
-    form_data.password = get_hashed_password(form_data.password)
+    form_data.password = authenticate.get_hashed_password(form_data.password)
     user = await user_crud.create_user(db, form_data)
-    access_token = create_access_token(user.username)
+    access_token = authenticate.create_access_token(user.username)
     return {"access_token": access_token, "token_type": "bearer"}
 
 @router.post("/token", response_model=user_schema.Token)
@@ -104,21 +24,21 @@ async def login_for_access_token(
     form_data: OAuth2PasswordRequestForm = Depends(),
     db: AsyncSession = Depends(get_db)
 ):
-    user = await authenticate_user(db, form_data.username, form_data.password)
+    user = await authenticate.authenticate_user(db, form_data.username, form_data.password)
     if not user:
         raise HTTPException(
             status_code=status.HTTP_401_UNAUTHORIZED,
             detail="Incorrect username or password",
             headers={"WWW-Authenticate": "Bearer"}
         )
-    access_token = create_access_token(user.username)
+    access_token = authenticate.create_access_token(user.username)
     return {"access_token": access_token, "token_type": "bearer"}
 
 @router.get("/users/me", response_model=user_schema.User)
-async def read_users_me(current_user: user_schema.User = Depends(get_current_active_user)):
+async def read_users_me(current_user: user_schema.User = Depends(authenticate.get_current_active_user)):
     return current_user
 
 @router.get("/users/me/tasks", response_model=List[task_schema.Task])
-async def read_own_items(current_user: user_schema.User = Depends(get_current_active_user)):
+async def read_own_items(current_user: user_schema.User = Depends(authenticate.get_current_active_user)):
     sample_task = task_schema.Task(id=1, title="my task", done=False)
     return [sample_task]
